Abstracted some controller functionality into a monitorsController and lcreated boilerplate for v2 events route
