+++ server/controllers/eventsController.ts
@@ -1,0 +1,12 @@
+
+
+interface EventsController {
+
+};
+
+const eventsController: EventsController = {
+
+  
+};
+
+export default eventsController;
