+++ server/controllers/keyspacesController.ts
@@ -7,34 +7,12 @@ import type { KeyspacesResponseBody, KeyspaceResponsePage, KeyDetails, Keyspace 
 import { getKeyspace } from './utils';
 
 interface KeyspacesController {
-  findAllMonitors: RequestHandler;
-  findSingleMonitor: RequestHandler;
   getKeyspacesForInstance: RequestHandler;
   refreshKeyspace: RequestHandler;
   getKeyspacePages: RequestHandler;
 }
 
 const keyspacesController: KeyspacesController = {
-
-  findAllMonitors: async (req, res, next) => {
-    res.locals.monitors = redisMonitors;
-    return next();
-  },
-
-  findSingleMonitor: async (req, res, next) => {
-
-    for (let monitor of redisMonitors) {
-      if (monitor.instanceId === +req.params.instanceId) {
-        res.locals.monitors = [monitor];
-      }
-    }
-
-    if (!res.locals.monitors) {
-      return next({ log: 'User provided invalid instanceId', status: 400, message: { err: 'Please provide a valid instanceId' } });
-    }
-
-    return next();
-  },
 
   getKeyspacesForInstance: async (req, res, next) => {
